backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0          # PostgreSQL async driver
psycopg2-binary==2.9.9   # Sync driver (Alembic)

# Cache
redis==5.0.4

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# HTTP + scraping
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.1
tenacity==8.3.0

# Data / validation
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

# Scheduler
apscheduler==3.10.4
